+++ braintree/braintree_gateway.py
@@ -8,6 +8,8 @@ from braintree.settlement_batch_summary_gateway import SettlementBatchSummaryGat
 from braintree.subscription_gateway import SubscriptionGateway
 from braintree.transaction_gateway import TransactionGateway
 from braintree.transparent_redirect_gateway import TransparentRedirectGateway
+from braintree.webhook_notification_gateway import WebhookNotificationGateway
+from braintree.webhook_testing_gateway import WebhookTestingGateway
 
 class BraintreeGateway(object):
     def __init__(self, config):
@@ -22,4 +24,5 @@ class BraintreeGateway(object):
         self.subscription = SubscriptionGateway(self)
         self.transaction = TransactionGateway(self)
         self.transparent_redirect = TransparentRedirectGateway(self)
-
+        self.webhook_notification = WebhookNotificationGateway(self)
+        self.webhook_testing = WebhookTestingGateway(self)