+++ braintree/util/http.py
@@ -1,6 +1,6 @@
-import httplib
 import base64
-from braintree.configuration import Configuration
+import braintree
+from braintree import version
 from braintree.util.xml_util import XmlUtil
 from braintree.exceptions.authentication_error import AuthenticationError
 from braintree.exceptions.authorization_error import AuthorizationError
@@ -9,7 +9,6 @@ from braintree.exceptions.not_found_error import NotFoundError
 from braintree.exceptions.server_error import ServerError
 from braintree.exceptions.unexpected_error import UnexpectedError
 from braintree.exceptions.upgrade_required_error import UpgradeRequiredError
-from braintree import version
 
 class Http(object):
     @staticmethod
@@ -50,31 +49,18 @@ class Http(object):
         return self.__http_do("PUT", path, params)
 
     def __http_do(self, http_verb, path, params=None):
-        if self.environment.is_ssl:
-            self.__verify_ssl()
-            conn = httplib.HTTPSConnection(self.environment.server, self.environment.port)
-        else:
-            conn = httplib.HTTPConnection(self.environment.server, self.environment.port)
-
-        conn.request(
-            http_verb,
-            self.config.base_merchant_path() + path,
-            XmlUtil.xml_from_dict(params) if params else '',
-            self.__headers()
-        )
-        response = conn.getresponse()
-        status = response.status
+        http_strategy = self.config.http_strategy()
+        request_body = XmlUtil.xml_from_dict(params) if params else ''
+        full_path = self.config.base_merchant_path() + path
+        status, response_body = http_strategy.http_do(http_verb, full_path, self.__headers(), request_body)
 
         if Http.is_error_status(status):
-            conn.close()
             Http.raise_exception_from_status(status)
         else:
-            data = response.read()
-            conn.close()
-            if len(data.strip()) == 0:
+            if len(response_body.strip()) == 0:
                 return {}
             else:
-                return XmlUtil.dict_from_xml(data)
+                return XmlUtil.dict_from_xml(response_body)
 
     def __authorization_header(self):
         return "Basic " + base64.encodestring(self.config.public_key + ":" + self.config.private_key).strip()
@@ -85,33 +71,6 @@ class Http(object):
             "Authorization": self.__authorization_header(),
             "Content-type": "application/xml",
             "User-Agent": "Braintree Python " + version.Version,
-            "X-ApiVersion": Configuration.api_version()
+            "X-ApiVersion": braintree.configuration.Configuration.api_version()
         }
 
-    def __verify_ssl(self):
-        if Configuration.use_unsafe_ssl: return
-
-        try:
-            import pycurl
-        except ImportError, e:
-            print "Cannot load PycURL.  Please refer to Braintree documentation."
-            print """
-If you are in an environment where you absolutely cannot load PycURL
-(such as Google App Engine), you can turn off SSL Verification by setting:
-
-    Configuration.use_unsafe_ssl = True
-
-This is highly discouraged, however, since it leaves you susceptible to
-man-in-the-middle attacks."""
-            raise e
-
-        curl = pycurl.Curl()
-        # see http://curl.haxx.se/libcurl/c/curl_easy_setopt.html for info on these options
-        curl.setopt(pycurl.CAINFO, self.environment.ssl_certificate)
-        curl.setopt(pycurl.ENCODING, 'gzip')
-        curl.setopt(pycurl.SSL_VERIFYPEER, 1)
-        curl.setopt(pycurl.SSL_VERIFYHOST, 2)
-        curl.setopt(pycurl.NOBODY, 1)
-        curl.setopt(pycurl.NOSIGNAL, 1)
-        curl.setopt(pycurl.URL, self.environment.protocol + self.environment.server_and_port)
-        curl.perform()