+++ tests/integration/test_credit_card.py
@@ -20,6 +20,7 @@ class TestCreditCard(unittest.TestCase):
         self.assertEquals("2009", credit_card.expiration_year)
         self.assertEquals("05/2009", credit_card.expiration_date)
         self.assertEquals("John Doe", credit_card.cardholder_name)
+        self.assertNotEquals(re.search("\A\w{32}\Z", credit_card.unique_number_identifier), None)
 
     def test_create_and_make_default(self):
         customer = Customer.create().customer