+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.14.0
+
+* Removes relative imports for python 3.0 (thanks [MichaelBlume](https://github.com/MichaelBlume))
+* Adds webhook gateways for parsing, verifying, and testing incoming notifications
+* Allow specifying the http strategy to use (PycURL, httplib, requests)
+
 ## 2.13.0
 
 * Adds search for duplicate credit cards given a payment method token